wizpix-ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.21
rembg==2.0.69
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
numpy==2.2.6
Pillow==12.0.0
requests==2.32.5
pytest==8.3.2
